requirements.txt
numpy==1.19.5
pyserial==3.5
